chore: reverted back to correct aostdio.c updated version for deployment
